Rename `regex-shorthand` as `better-regex`
